+++ concert-reservation-application/src/main/java/concert/application/ConcertApplication.java
@@ -1,17 +1,18 @@
 package concert.application;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@SpringBootApplication(scanBasePackages="concert")
-@EnableScheduling
-@EnableAspectJAutoProxy(proxyTargetClass = true)
+@SpringBootApplication
+@ComponentScan(basePackages = {"concert-reservation-domain", "concert-reservation-infrastructure"})
 public class ConcertApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(ConcertApplication.class, args);
 	}
-
 }
