--- concert-reservation-application/src/main/java/concert/application/reservation/infrastructure/kafka/KafkaMessageProducer.java
@@ -1,35 +1,0 @@
-package concert.application.reservation.infrastructure.kafka;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import concert.application.reservation.application.event.PaymentConfirmedEvent;
-import concert.application.reservation.application.event.PaymentRequestEvent;
-import lombok.RequiredArgsConstructor;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class KafkaMessageProducer {
-
-  private final KafkaTemplate<String, String> kafkaTemplate;
-  private final ObjectMapper objectMapper;
-
-  public void sendPaymentRequestEvent(String topic, PaymentRequestEvent event) {
-    try {
-      String eventJson = objectMapper.writeValueAsString(event);
-      kafkaTemplate.send(topic, eventJson);
-    } catch (JsonProcessingException e) {
-      throw new RuntimeException("Failed to serialize event to JSON", e);
-    }
-  }
-
-  public void sendPaymentConfirmedEvent(String topic, PaymentConfirmedEvent event) {
-    try {
-      String eventJson = objectMapper.writeValueAsString(event);
-      kafkaTemplate.send(topic, eventJson);
-    } catch (JsonProcessingException e) {
-      throw new RuntimeException("Failed to serialize event to JSON", e);
-    }
-  }
-}
