+++ concert-reservation-domain/src/main/java/concert/domain/concert/cache/ConcertCache.java
@@ -1,12 +1,11 @@
 package concert.domain.concert.cache;
 
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import concert.domain.concert.domain.Concert;
 
 import java.util.List;
 
 public interface ConcertCache {
-    void saveTop30Concerts(List<Concert> concerts) throws JsonProcessingException;
+    void saveTop30Concerts(List<Concert> concerts) throws JsonProcessingException, JsonProcessingException;
     List<Concert> findTop30Concerts() throws JsonProcessingException;
 }
