+++ concert-reservation-domain/src/main/java/concert/domain/waitingQueue/domain/WaitingQueueDao.java
@@ -1,0 +1,15 @@
+package concert.domain.waitingQueue.domain;
+
+public interface WaitingQueueDao {
+
+    String addToWaitingQueue(long concertId, String uuid);
+
+    long getWaitingRank(long concertId, String uuid);
+
+    String getActiveQueueToken(long concertId, String uuid);
+
+    void deleteActiveQueueToken(long concertId, String uuid);
+
+    void removeTop333FromWaitingQueue(long concertId);
+
+}
