+++ concert-reservation-infrastructure/src/main/java/redis/RedisConcertCache.java
@@ -1,4 +1,4 @@
-package concert.domain.redis;
+package redis;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -6,22 +6,24 @@ import concert.domain.concert.cache.ConcertCache;
 import concert.domain.concert.domain.Concert;
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.redisson.api.RMap;
 import org.redisson.api.RedissonClient;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
 import java.util.List;
+import java.util.logging.Logger;
 
-@Slf4j
 @Primary
 @Component
 @RequiredArgsConstructor
 public class RedisConcertCache implements ConcertCache {
 
     private final RedissonClient redisson;
+
+    private static final Logger log = (Logger) LoggerFactory.getLogger(RedisConcertCache.class);
     private static final String TOP30_CONCERTS = "top30concerts";
     private final ObjectMapper objectMapper;
 
@@ -46,18 +48,18 @@ public class RedisConcertCache implements ConcertCache {
             log.info("Retrieved top 30 concerts from Redis.");
             return objectMapper.readValue(top30concertListJson, objectMapper.getTypeFactory().constructCollectionType(List.class, Concert.class));
         } else {
-            log.warn("No top 30 concerts found in Redis.");
+            log.warning("No top 30 concerts found in Redis.");
             return null;
         }
     }
 
     // Fallback 메소드 정의 (서킷 브레이커가 열렸을 때 호출됨)
     public void fallbackSaveConcerts(String concerts, Throwable t) {
-        log.error("Failed to save concerts to Redis. Circuit breaker is open.", t);
+        log.severe("Failed to save concerts to Redis. Circuit breaker is open.");
     }
 
     public String fallbackGetConcerts(Throwable t) {
-        log.error("Failed to retrieve concerts from Redis. Circuit breaker is open.", t);
+        log.severe("Failed to retrieve concerts from Redis. Circuit breaker is open.");
         return null;
     }
 }
